audio: fix earc Cert HFR5-1-35 failed [1/1]
PD#SWPL-159771

Problem:
get CDS info which is not ready by the spec.
rootcause:when the earc connection success, it send
the earc state to framework, then send connet msg to
audio hal, and audio hal get cds info, but the cds is
not ready by the spec.

Solution:
when we get the cds ready info, then send the earc
connection success event.

Verify:
T3X

Change-Id: Id48a83fd20836fb1e61e1f0f9330bd3b6d6515a6
Signed-off-by: jian zhou <jian.zhou@amlogic.com>