ASoC: rockchip: pdm_v2: set the pdm default gain to 24dB
Set the default gain 24dB, this parameter can get better
performance if the voice energy is lower. In other words this
can improve PDM IP SNR.

So the applicable range of this is for sound intensity below 100dB.
If you want to record stronger sound intensity, you must set
PDM gain register but not soft gain-controller.

Change-Id: Ie8883064d124745c8ac2eec2034ccf94dcd362cb
Signed-off-by: Jason Zhu <jason.zhu@rock-chips.com>