arm64: dts: rockchip: Add OP-TEE firmware node for rk3576-linux.dtsi
Add a new firmware node for OP-TEE to the rk3576-linux.dtsi file. OP-TEE is a
trusted execution environment that provides secure execution of trusted
applications.

The firmware node includes the following properties:

- compatible: Set to "linaro,optee-tz" to identify the node as an OP-TEE
  firmware node.
- method: Set to "smc" to specify that the OP-TEE firmware is accessed using
  the ARM Secure Monitor Call (SMC) instruction.

This change allows the kernel to properly initialize and use the OP-TEE
firmware on the RK3576 platform, enabling secure execution of trusted
applications and providing additional security features.

Change-Id: Id8c2fcfcb9176c90afe82d12ee9392e20c8446d1
Signed-off-by: Caesar Wang <wxt@rock-chips.com>