drm/rockchip: dsi: add support PSR for mipi command mode
As described in:
https://patchwork.freedesktop.org/patch/msgid/20190228210939.83386-2-sean@poorly.run

From the driver's perspective, this works like a regular disable/enable
cycle. The driver need only check the 'psr_transition' state in
connector_state and keep the panel turned on when in .disable(), while
everything else will cycle off as normal. If drivers want more control,
they can use the psr_transition state to enter a low-power state to
minimize PSR exit time.

While this carries the PSR moniker, it is not specific to the
DisplayPort technology. This can be used for power savings
with other types of self refresh, such as MIPI command mode.

Change-Id: I8009e1e369d7beeef900b96235b6fb1545a5f0eb
Signed-off-by: Guochun Huang <hero.huang@rock-chips.com>