arm64: rockchip_linux_defconfig: Disable CPU errata 2441007 (Cortex-A55)
According to ANDROID commit a9567a35d0b8 ("ANDROID: arm64: Disable workaround
for CPU errata 2441007 and 2441009"):

CPU errata 2441007 (Cortex-A55) and 2441009 (Cortex-A510) are categorised
as "rare" by Arm and consequently the workaround is not intended to be
deployed in practice as the issue is not expected to occur in real-world
environments.

Given that the cost of the workaround, which issues additional broadcast
TLB invalidation requests, has been shown to impact kswapd significantly
on Pixel devices, disable the workaround following Arm's recommendation.

Signed-off-by: Tao Huang <huangtao@rock-chips.com>
Change-Id: I794d048a094714c1370752bfaa47bee2afd57d53