phy: rockchip: inno-usb2: add usb2 phy support for rk3506
rk3506 SOC has two versions, which are rk3506g and rk3506b. They both
have two otg ports and uses one USB PHY with two ports. The OTG0 port
support BC1.2 detect.

They are different in that rk3506g don't have vbus and id detect pin
and use gpio to replace it. We use the extcon-usb-gpio.c driver to
manage the interrupts. If we get the EXTCON_USB state, it means that
the vbus is high and iddig is high, we enable the bvalid_phy_con to
trigger bvalid interrupt.

Change-Id: Iac0d0a992ee2d2f9664bea36d06f6b060a5bfb99
Signed-off-by: Jianwei Zheng <jianwei.zheng@rock-chips.com>