phy: rockchip: inno-usb2: Support GPIO ID detect for USB2PHY
This patch register gpio id notifier to support USB2PHY to get extcon
message from extcon-usb-gpio.c driver. The extcon-gpio-usb.c driver
returns the state based on the ID and Vbus pin values as shown below.

      State              |    ID   |   VBUS
     ----------------------------------------
      [1] USB            |    H    |    H
      [2] none           |    H    |    L
      [3] USB-HOST       |    L    |    H
      [4] USB-HOST       |    L    |    L

There is no need to control usb bvalid when USB_HOST state is true.
When USB_HOST state is true, we need to configure the iddig related
registers to trigger the controller's ID interrupt and set the
controller to HOST mode. When USB_HOST state is false, we need to
restore the register configuration.

Change-Id: Ia17fa67f5a26b2e5d989ede23ee6243cdc52f05f
Signed-off-by: Jianwei Zheng <jianwei.zheng@rock-chips.com>
Signed-off-by: William Wu <william.wu@rock-chips.com>