ASoC: rockchip: pdm_v2: Fix clk glitch on runtime PM
For controller which is managed by PD (power-domain),
when PD off, the controller is reset to the default
status, and the FRAC-DIV is a fixed value(1/20).

Once the mclk is enabled, there are some high freq cycle
leak, to fix this issue, we use the pinctrl-idle to
block these cycles until the config has been come back
to the normal state.

Ref: commit 1f8e86a5ea59 ("ASoC: rockchip: pdm: Fix clk glitch on runtime PM")

Signed-off-by: XiaoTan Luo <lxt@rock-chips.com>
Change-Id: I8b1226896d8ca20293335a879452df732801f712