drm/rockchip: tve: add .loader_protect() support for tve
If logo display enabled, the rockchip_tve_encoder_loader_protect()
helps to set the tve enable/disable status and call
pm_runtime_get_sync()/pm_runtime_put(), which allows for
the continuation of the display state from the U-Boot stage.

In addition, fix the check of tve mode setting in .mode_set()
to avoid display flashing, which caused by the off/on switch of dac.

Change-Id: I6f1fcaf05cb6e362ce375e447c566fb5356c5380
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>