ASoC: codecs: rk_dsm: add support iomux switching for anti-pop automatically
The LP/LN and RP/RN differential pairs of rkdsm should keep
in the same direction as much as possible when working and
shutting down, otherwise the common mode output is likely
to cause pop sound.

Therefore, we can introduce the ioc register address as REG
resource in dts, and switch the dsm pin to function and io
state respectively when dsm starts and stops.

Change-Id: I4a3e6f5e45f030fdbab74ee7d9a769fc8d77f86d
Signed-off-by: Xing Zheng <zhengxing@rock-chips.com>