PCI: dw: rockchip: Add shutdown support
There is nothing worth to do when rebooting in normal system
as the whole thing is going to be reset in a clean way. However
if the reboot process doesn't reset the anything, then reset
PCIe controller in probe would throw a glitch which could generate
a spurious irq. Work around this by disabling misc irq before rebooting.

Signed-off-by: Shawn Lin <shawn.lin@rock-chips.com>
Change-Id: I7b6e53ecbce385581a6beb0333a2df320a32cc9e