Revert "PCI: rockchip: dw: remove wakeup if attached device is down"
This reverts commit e15707617c807e82a706dfdd2240c83d62f2634e.

This need the system to rmmod wireless ko before suspend, otherwise even if
the power is down but removing wakeup will call the sub-device driver to
broadcast this uevent to the userspace which prevent the system from into
suspend. Whether to rmmod wireless driver is system depend, so revert this
commit.

Signed-off-by: Shawn Lin <shawn.lin@rock-chips.com>
Change-Id: I9f0467a7403c1f18820a89ae4707d3bd469910a6