drm/rockchip: analogix_dp: fix NULL pointer issue when disable encoder
When disable encoder, the encoder.crtc will be reference. In some
case, for example show logo stage, if some config cause a full
modeset, the encoder.crtc will be used before assigned a value,
which will cause a NUll pointer issue.
It's better to use drm_connector_state.crtc to avoid this issue.

Change-Id: I0382eda8a7e25e15b06cefc08ec7be94cf8136ea
Signed-off-by: Zhang Yubing <yubing.zhang@rock-chips.com>