audio: resample num error [1/1] for fix bt mic cannot work bug
PD#SWPL-169776

Problem:
chip info num is not meeting with the real used resample numble.
resample check error when resample B is disabled.
dts only enable resample A, when check resample number, the
resample attach module will error. For example, chip info numble set to
2;when resample A and B all are enabled, the attach module will be 8
for hdmi in and 7 for loopback A; if disable  resample B, the attach
module will be 8 and 0. once tinycap tdmin with to->src = 0, the toddr
bit 27 will be enable, which control the no use bit and cause toddr
hangup

Solution:
1. fix check resample number flow, by use real probe chip info and id

Verify:
ross

Change-Id: I4076024995fd5fdbf5dbbb6b2d7effa990cfbf8e
Signed-off-by: qing.zhang <qing.zhang@amlogic.com>