drm/rockchip: bridge: analogix_dp: fix the regs settings for TEST_PATTERN_80BIT
According to the DP PHY CTS v1.2b 3.2.3 and 3.3.3 chapters, the
PreEmphasis Level Test Pattern(PLTPAT), which also was called
80-bit custom pattern, should be:

1111100000 1111100000 1111100000 1111100000
1111100000 1111100000 1111100000 1111100000

It is used to verify the HBR2 in "Non Pre-Emphasis Level Verification
Testing" and "Pre-Emphasis Level Verification and Maximum Pk-Pk
Differential Voltage Testing".

Without this patch, the PLTPAT pattern is incorrect.

Fixes: 304dbb104da1 ("drm/bridge: analogix_dp: Add DP Test Automation")
Change-Id: I4b00acd0a360a8ba3f3eaf32fb30e0a18a451f5c
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>